move files; final form logic
