+++ packages/web/src/pages/upload-page/components/TrackModalArray.tsx
@@ -1,8 +1,8 @@
 import { ReleaseDateModalForm } from '../fields/ReleaseDateModalForm'
 import { RemixModalForm } from '../fields/RemixModalForm'
 import { SourceFilesModalForm } from '../fields/SourceFilesModalForm'
+import { TrackAvailabilityModalForm } from '../fields/TrackAvailabilityModalForm'
 
-import { TrackAvailabilityModalForm } from './TrackAvailabilityModalForm'
 import styles from './TrackModalArray.module.css'
 
 export const TrackModalArray = () => {
