+++ packages/web/src/pages/upload-page/fields/CollectibleGatedFields.module.css
@@ -1,0 +1,56 @@
+.root {
+  padding-top: var(--unit-2);
+}
+
+.dropdown {
+  z-index: 10000;
+}
+
+.root .dropdownInput:has(:global(.ant-select)) {
+  height: 72px;
+}
+
+.dropdownInput :global(.ant-select-selection-item),
+.dropdownInput :global(.ant-select-selection-placeholder) > div {
+  display: flex;
+  align-items: center;
+}
+
+.root .dropdown :global(.ant-select-item) {
+  padding: 16px;
+  height: 80px;
+}
+.dropdownRow {
+  display: flex;
+  align-items: center;
+  z-index: 3000;
+}
+.dropdownRow > span {
+  margin-left: 8px;
+  font-weight: var(--font-demi-bold);
+  font-size: var(--font-l);
+}
+
+.dropdownInput :global(.ant-select-selection-item) {
+  pointer-events: none;
+}
+
+:global(.ant-select-selection-item) .dropdownRow > img {
+  width: 32px;
+  height: 32px;
+  border-radius: 8px;
+}
+
+:global(.ant-select-item) .dropdownRow > img {
+  width: 48px;
+  height: 48px;
+  border-radius: 8px;
+}
+
+.helpCallout {
+  border-right: none;
+  border-left: none;
+  border-bottom: none;
+  border-radius: initial;
+  height: 100%;
+}
