+++ packages/web/src/pages/upload-page/fields/HiddenAvailabilityFields.module.css
